pyproject.toml
[tool.poetry]
name = "flexpart-ifs-preprocessor"
version = "0.2.0"
description = "Preprocesses IFS data for use in Flexpart"
license = "BSD-3-Clause"
authors = [
    "Nina Burgdorfer <nina.burgdorfer@meteoswiss.ch>",
    "Victoria Cherkas <victoria.cherkas@meteoswiss.ch>"
    ]
readme = "README.rst"
include = ["HISTORY.rst"]

[[tool.poetry.source]]
name = "meteoswiss"
url = "https://service.meteoswiss.ch/nexus/repository/python-all/simple"
priority = "primary"

[tool.poetry.dependencies]
python = "~3.11"
click = "^8.2.1"
boto3 = "^1.34.115"
jinja2 = ">=3.1.1"
numpy = "^1.26.4"
xarray = ">=2023.4.1"
eccodes = ">=1.7"
flexprep = "^0.2.1"
mchpy = "^4.0.4"
aws-lambda-powertools = "^3.24.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
pre-commit = "^3.7.1"
codespell = "^2.3.0"
flake8 = "^7.1.0"
pytest-cov = "*"
pylint = "*"
yapf = "*"
mypy = "*"
sphinx = "*"
importlib_resources = "*"
pydata-sphinx-theme = "*"
sphinx-autoapi = "*"
moto = "^5.1.22"

[tool.poetry.scripts]
flexpart-ifs-preprocessor = "flexpart_ifs_preprocessor.main:cli"
[tool.yapf]
based_on_style = "pep8"
column_limit = "120"

[tool.pylint.master]
disable = [
    'C0114', # missing-module-docstring
    'C0115', # missing-class-docstring
    'C0116', # missing-function-docstring
    'W0511', # fix me or to-do comments are already covered by SonarQube
]

[tool.pylint.basic]
argument-naming-style = 'any'
attr-naming-style = 'any'

[tool.pylint.format]
# Maximum number of lines in a module.
max-line-length = 120

[tool.pylint.design]
# Minimum number of public methods for a class (see R0903).
min-public-methods = 0

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
omit = ["test/**",  "doc/**"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
